+++ src/Lib/Vtk/Algorithm/PointSetAlgorithm/quickVtkWarpTo.hpp
@@ -12,7 +12,7 @@ namespace quick {
         class WarpTo : public PointSetAlgorithm {
             Q_OBJECT
             Q_PROPERTY(quick::Math::Vector3* position READ getPosition CONSTANT);
-            Q_PROPERTY(float scaleFactor READ getScaleFactor WRITE setScaleFactor NOTIFY scaleFactorChanged);
+            Q_PROPERTY(qreal scaleFactor READ getScaleFactor WRITE setScaleFactor NOTIFY scaleFactorChanged);
             Q_PROPERTY(bool absolute READ getAbsolute WRITE setAbsolute NOTIFY absoluteChanged);
         private:
             static Qml::Register::Symbol::Class<WarpTo> Register;
@@ -21,8 +21,8 @@ namespace quick {
         public:
             WarpTo();
             auto getPosition() -> Math::Vector3*;
-            auto setScaleFactor(float) -> void;
-            auto getScaleFactor() -> float;
+            auto setScaleFactor(qreal) -> void;
+            auto getScaleFactor() -> qreal;
             auto setAbsolute(bool) -> void;
             auto getAbsolute() -> bool;
         signals:
