+++ src/Lib/Vtk/Algorithm/PolyDataAlgorithm/quickVtkHedgeHog.hpp
@@ -16,15 +16,15 @@ namespace quick {
                 UseNormal = VTK_USE_NORMAL
             };
             Q_ENUM(VectorMode);
-            Q_PROPERTY(float scaleFactor READ getScaleFactor WRITE setScaleFactor NOTIFY scaleFactorChanged);
+            Q_PROPERTY(qreal scaleFactor READ getScaleFactor WRITE setScaleFactor NOTIFY scaleFactorChanged);
             Q_PROPERTY(VectorMode vectorMode READ getVectorMode WRITE setVectorMode NOTIFY vectorModeChanged);
         private:
             static Qml::Register::Symbol::Class<HedgeHog> Register;
             vtkSmartPointer<vtkHedgeHog> m_vtkObject;
         public:
             HedgeHog();
-            auto setScaleFactor(float) -> void;
-            auto getScaleFactor() -> float;
+            auto setScaleFactor(qreal) -> void;
+            auto getScaleFactor() -> qreal;
             auto setVectorMode(VectorMode) -> void;
             auto getVectorMode() -> VectorMode;
         signals:
