+++ src/Lib/Vtk/ImplicitFunction/quickVtkSphere.hpp
@@ -11,7 +11,7 @@ namespace quick {
 
         class Sphere : public ImplicitFunction {
             Q_OBJECT
-            Q_PROPERTY(float radius READ getRadius WRITE setRadius NOTIFY radiusChanged);
+            Q_PROPERTY(qreal radius READ getRadius WRITE setRadius NOTIFY radiusChanged);
             Q_PROPERTY(quick::Math::Vector3* center READ getCenter CONSTANT);
         private:
             static Qml::Register::Symbol::Class<Sphere> Register;
@@ -19,12 +19,12 @@ namespace quick {
             Math::Vector3* m_center = nullptr;
         public:
             Sphere();
-            auto setRadius(float) -> void;
-            auto getRadius() -> float;
+            auto setRadius(qreal) -> void;
+            auto getRadius() -> qreal;
             auto setCenter(Math::Vector3*) -> void;
             auto getCenter() -> Math::Vector3*;
         public slots:
-            float evaluateFunction(float, float, float);
+            qreal evaluateFunction(qreal, qreal, qreal);
         signals:
             void radiusChanged();
         };
