+++ src/Lib/Vtk/Mapper/AbstractMapper3D/AbstractVolumeMapper/VolumeMapper/quickVtkSmartVolumeMapper.cpp
@@ -10,23 +10,23 @@ namespace quick {
             this->m_vtkObject = vtkSmartVolumeMapper::SafeDownCast(this->Algorithm::getVtkObject());
         }
 
-        auto SmartVolumeMapper::setFinalColorWindow(float finalColorWindow) -> void {
+        auto SmartVolumeMapper::setFinalColorWindow(qreal finalColorWindow) -> void {
             this->m_vtkObject->SetFinalColorWindow(finalColorWindow);
             emit this->finalColorWindowChanged();
             this->update();
         }
 
-        auto SmartVolumeMapper::getFinalColorWindow() -> float {
+        auto SmartVolumeMapper::getFinalColorWindow() -> qreal {
             return this->m_vtkObject->GetFinalColorWindow();
         }
 
-        auto SmartVolumeMapper::setFinalColorLevel(float finalColorLevel) -> void {
+        auto SmartVolumeMapper::setFinalColorLevel(qreal finalColorLevel) -> void {
             this->m_vtkObject->SetFinalColorLevel(finalColorLevel);
             emit this->finalColorLevelChanged();
             this->update();
         }
 
-        auto SmartVolumeMapper::getFinalColorLevel() -> float {
+        auto SmartVolumeMapper::getFinalColorLevel() -> qreal {
             return this->m_vtkObject->GetFinalColorLevel();
         }
     }
